LayoutTests/fast/workers/worker-messageport-gc-expected.txt

CONSOLE MESSAGE: line 18: ReferenceError: Can't find variable: MessageChannel
FAIL: Timed out waiting for notifyDone to be called
Test that workers stay reachable via message ports. Should print "DONE" when done.
